requirements.txt
streamlit
openai
unstructured
chromadb
pillow
python-dotenv
